refactor: Consolidate API lib modules under _lib directory
Move shared utility modules (store, license-key, email, plan-config) from
api/lib and api/ root into api/_lib for consistent organization. Update all
import paths accordingly and format for readability.
