+++ api/_lib/license-key.js
